fix(#14): Wrong escaping of double quotes when they are contained within single-quoted strings (either '' or NOWDOC)
Single-quoted strings were being escaped to double quotes through an existing intention. This was not bad per-se, but had the unintended faulty consequence that the double quotes contained within a single-quoted strings would be escaped.
Escaping double quotes is not necessary in HEREDOC. Having them "escaped" causes HEREDOC to print additional backslashes, causing the end result to be different from the starting concatenation.